src/tdd_1/SqliteDBMaker.java
package tdd_1;

import java.sql.Connection;
import java.sql.DatabaseMetaData;
import java.sql.DriverManager;
import java.sql.SQLException;

public class SqliteDBMaker {
	public static void createNewDatabase(String fileName) {
		 
        String url = "jdbc:sqlite:/home/jjcombs/workspace-fcdtdebug/" + fileName;
 
        try (Connection conn = DriverManager.getConnection(url)) {
            if (conn != null) {
                DatabaseMetaData meta = conn.getMetaData();
                System.out.println("The driver being used is called " + meta.getDriverName());
                System.out.println("A new database has been created.");
            }
 
        } catch (SQLException e) {
            System.out.println(e.getMessage());
        }
    }
}
